+++ src/main/java/com/chicchoc/sivillage/global/auth/application/AuthService.java
@@ -1,8 +1,8 @@
 package com.chicchoc.sivillage.global.auth.application;
 
-import com.chicchoc.sivillage.global.auth.dto.SignInRequestDto;
-import com.chicchoc.sivillage.global.auth.dto.SignInResponseDto;
-import com.chicchoc.sivillage.global.auth.dto.SignUpRequestDto;
+import com.chicchoc.sivillage.global.auth.dto.in.SignInRequestDto;
+import com.chicchoc.sivillage.global.auth.dto.in.SignUpRequestDto;
+import com.chicchoc.sivillage.global.auth.dto.out.SignInResponseDto;
 
 public interface AuthService {
