+++ src/main/java/com/chicchoc/sivillage/global/jwt/presentation/TokenController.java
@@ -1,9 +1,9 @@
-package com.chicchoc.sivillage.global.auth.presentation;
+package com.chicchoc.sivillage.global.jwt.presentation;
 
-import com.chicchoc.sivillage.global.auth.application.TokenService;
-import com.chicchoc.sivillage.global.auth.dto.CreateAccessTokenRequestDto;
-import com.chicchoc.sivillage.global.auth.dto.CreateAccessTokenResponseDto;
-import com.chicchoc.sivillage.global.auth.jwt.JwtProperties;
+import com.chicchoc.sivillage.global.jwt.application.TokenService;
+import com.chicchoc.sivillage.global.jwt.dto.in.CreateAccessTokenRequestDto;
+import com.chicchoc.sivillage.global.jwt.dto.out.CreateAccessTokenResponseDto;
+import com.chicchoc.sivillage.global.jwt.config.JwtProperties;
 import com.chicchoc.sivillage.global.common.entity.CommonResponseEntity;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
