+++ src/test/java/com/chicchoc/sivillage/global/config/jwt/TokenProviderTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import com.chicchoc.sivillage.domain.member.domain.Gender;
 import com.chicchoc.sivillage.domain.member.domain.Member;
 import com.chicchoc.sivillage.domain.member.infrastructure.MemberRepository;
-import com.chicchoc.sivillage.global.auth.jwt.JwtProperties;
-import com.chicchoc.sivillage.global.auth.jwt.JwtTokenProvider;
+import com.chicchoc.sivillage.global.jwt.config.JwtProperties;
+import com.chicchoc.sivillage.global.jwt.application.JwtTokenProvider;
 import com.chicchoc.sivillage.global.common.generator.NanoIdGenerator;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
